task-management-project/job_app/serializers.py
from rest_framework import serializers
from job_app.models import Job


class JobSerializer(serializers.ModelSerializer):
    class Meta:
        model = Job
        fields = ['id', 'title', 'company', 'location', 'description', 'salary', 'created_at']
