task-management-project/requirements.txt
django
django-cors-headers
djangorestframework
django-environ
psycopg2-binary